[MultiverseAction] Added multiverse demo.
